Update lab 3
