--- Lab_3/main.c
@@ -1,71 +1,0 @@
-/*
-    Author:     Jesus Minjares
-                Master of Science in Computer Engineering
-    Course:     EE 5190 Laboratory for Microprocessors Systems II        
-    Lab 3:
-                Understand how to use the FreeRTOS with Espressif.
-            Tasks
-                Task 1 -> toggle `onboard led`
-                Task 2 -> print `hello world` every second
-            Undergrad Bonus:
-                Create an other task that will toggle the led and print it state in the terminal
-            Grad Bonus:
-                Create addtional task of to run any led sequence of the previous labs and it should run every 5 seconds
-    Date:       09-05-2021
-
-    Pinout:
-
-                                         +-----------------------+
-                                         | O      | USB |      O |
-                                         |        -------        |
-                                     3V3 | [ ]               [ ] | VIN
-                                     GND | [ ]               [ ] | GND
-     Touch3 / HSPI_CS0 / ADC2_3 / GPIO15 | [ ]               [ ] | GPIO13 / ADC2_4 / HSPI_ID / Touch4
- CS / Touch2 / HSPI_WP / ADC2_2 /  GPIO2 | [ ]               [ ] | GPIO12 / ADC2_5 / HSPI_Q / Touch5
-      Touch0 / HSPI_HD / ADC2_0 /  GPIO4 | [ ]               [ ] | GPIO14 / ADC2_6 / HSPI_CLK / Touch6
-                         U2_RXD / GPIO16 | [ ]               [ ] | GPIO27 / ADC2_7 / Touch7
-                         U2_TXD / GPIO17 | [ ]               [ ] | GPIO26 / ADC2_9 / DAC2
-                      V_SPI_CS0 /  GPIO5 | [ ]  ___________  [ ] | GPIO25 / ADC2_8 / DAC1
-                SCK / V_SPI_CLK / GPIO18 | [ ] |           | [ ] | GPIO33 / ADC1_5 / Touch8 / XTAL32
-        U0_CTS / MSIO / V_SPI_Q / GPIO19 | [ ] |           | [ ] | GPIO32 / ADC1_4 / Touch9 / XTAL32
-                 SDA / V_SPI_HD / GPIO21 | [ ] |           | [ ] | GPIO35 / ADC1_7 
-                  CLK2 / U0_RXD /  GPIO3 | [ ] |           | [ ] | GPIO34 / ADC1_6 
-                  CLK3 / U0_TXD /  GPIO1 | [ ] |           | [ ] | GPIO39 / ADC1_3 / SensVN 
-        SCL / U0_RTS / V_SPI_WP / GPIO22 | [ ] |           | [ ] | GPIO36 / ADC1_0 / SensVP 
-                MOSI / V_SPI_WP / GPIO23 | [ ] |___________| [ ] | EN 
-                                         |                       |
-                                         |  |  |  ____  ____  |  |
-                                         |  |  |  |  |  |  |  |  |
-                                         |  |__|__|  |__|  |__|  |
-                                         | O                   O |
-                                         +-----------------------+
-*/
-#include <stdio.h>
-#include "freertos/FreeRTOS.h"
-#include "freertos/task.h"
-#include "esp_system.h"
-#include "driver/gpio.h"
-/* Define pin 13 as a “BLINK_GPIO” */
-#define BLINK_GPIO 13
-/* Code for the hello_task task */
-void hello_task(void *pvParameter)
-{
-    
-}
-/* Code for the blinky task */
-void blinky_task(void *pvParameter)
-{
-    
-}
-void led_sequence_task(void *pvParameter){
-
-}
-void app_main()
-{
-    /* Create the task defined by xTaskCreate.*/
-    xTaskCreate(&hello_task, "hello_task", 2048, NULL, 5, NULL);
-    xTaskCreate(&blinky, "blinky", 512, NULL, 5, NULL);
-
-    /*Create any additional tasks*/
-    
-}
